Fix an issue in the event code (from rhuidean).
Allow the same event to be posted more than once to the same queue. If
this is done it can allow things like two "this socket is ready to be
read from" events to be posted and result in two calls to read(), one
of which will wind up blocking. This is bad. However, NOT allowing this
results in worse things, like if a certain event-triggering data comes
in more than once in a single read()/event loop it results it it only
being processed once and will ignore the rest. This is unacceptable.
Since the former can be corrected with good coding, and the latter
cannot, then the former loses (thanks sycobuny).