Fix things broken by new config, try to do support for 1.8. Cleanups.
Make an attempt at supporting 1.8. Unfortunately, `IO::WaitReadable` is
still holding us up on that. Spread the DSL config changes throughout
the rest of the app and a connecting client now works as far as it did
before I started messing with things.

Need to work on real tests. Bacon or riot?